+++ packages/suite/src/storage/migrations/versions/26.4.0.ts
@@ -2,6 +2,6 @@ import { createMigration } from '@suite/idb-migration-utils';
 
 import { SuiteDBSchema } from 'src/storage/definitions';
 
-export default createMigration<SuiteDBSchema>('26.3.0', db => {
-    db.createObjectStore('suiteSyncOwners');
+export default createMigration<SuiteDBSchema>('26.4.0', db => {
+    db.createObjectStore('phishing');
 });